+++ PTSLAttendanceManager/Models/Entity/UserWiseLeave.cs
@@ -5,10 +5,10 @@
         public long Id { get; set; }
         public required string UserId { get; set; }
         public required Users User { get; set; }
-        public long PendingEarnedLeave { get; set; }
-        public long PendingCasualLeave { get; set; }
-        public long PendingSickLeave { get; set; }
-        public long PendingMaternityLeave { get; set; }
+        public float PendingEarnedLeave { get; set; }
+        public float PendingCasualLeave { get; set; }
+        public float PendingSickLeave { get; set; }
+        public float PendingMaternityLeave { get; set; }
         public bool IsActive { get; set; } = true;
